+++ torch/onnx/_internal/torchscript_exporter/symbolic_opset12.py
@@ -7,11 +7,14 @@ import sys
 
 import torch
 from torch._C import _onnx as _C_onnx
-from torch.onnx import errors, symbolic_helper, symbolic_opset9 as opset9, utils
+from torch.onnx import errors
 from torch.onnx._internal.torchscript_exporter import (
     _type_utils,
     jit_utils,
     registration,
+    symbolic_helper,
+    symbolic_opset9 as opset9,
+    utils,
 )
