+++ torch/onnx/_internal/torchscript_exporter/symbolic_opset8.py
@@ -36,11 +36,13 @@ import warnings
 
 import torch
 from torch._C import _onnx as _C_onnx
-from torch.onnx import errors, symbolic_helper, symbolic_opset9 as opset9
+from torch.onnx import errors
 from torch.onnx._internal.torchscript_exporter import (
     _type_utils,
     jit_utils,
     registration,
+    symbolic_helper,
+    symbolic_opset9 as opset9,
 )
